requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0

sqlalchemy==2.0.36
pydantic==2.9.2
python-multipart==0.0.12
anyio==4.11.0
typing-extensions==4.12.2

python-jose==3.3.0
passlib[bcrypt]==1.7.4
requests==2.32.3

chromadb==0.4.24
sentence-transformers==3.2.1

pdfplumber==0.11.3
python-docx==1.1.2
PyPDF2==3.0.1
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0

openai==1.52.2
tqdm==4.67.1

streamlit==1.39.0
pytest==8.3.3
